+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/StockLocationLineServiceSupplychainImpl.java
@@ -20,6 +20,8 @@ package com.axelor.apps.supplychain.service;
 import com.axelor.apps.base.db.Product;
 import com.axelor.apps.stock.db.StockLocation;
 import com.axelor.apps.stock.db.StockLocationLine;
+import com.axelor.apps.stock.db.TrackingNumber;
+import com.axelor.apps.stock.service.StockLocationLineService;
 import com.axelor.apps.stock.service.StockLocationLineServiceImpl;
 import com.axelor.apps.supplychain.exception.IExceptionMessage;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
@@ -27,11 +29,14 @@ import com.axelor.exception.AxelorException;
 import com.axelor.exception.db.repo.TraceBackRepository;
 import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
+import com.google.inject.Inject;
 import com.google.inject.servlet.RequestScoped;
 import java.math.BigDecimal;
 
 @RequestScoped
 public class StockLocationLineServiceSupplychainImpl extends StockLocationLineServiceImpl {
+
+  @Inject protected AppSupplychainService appSupplychainService;
 
   @Override
   public void checkIfEnoughStock(StockLocation stockLocation, Product product, BigDecimal qty)
@@ -73,6 +78,30 @@ public class StockLocationLineServiceSupplychainImpl extends StockLocationLineSe
     BigDecimal availableQty = BigDecimal.ZERO;
     if (stockLocationLine != null) {
       availableQty = stockLocationLine.getCurrentQty().subtract(stockLocationLine.getReservedQty());
+    }
+    return availableQty;
+  }
+
+  @Override
+  public BigDecimal getTrackingNumberAvailableQty(
+      StockLocation stockLocation, TrackingNumber trackingNumber) {
+
+    if (!appSupplychainService.isApp("supplychain")
+        || !appSupplychainService.getAppSupplychain().getManageStockReservation()) {
+      return super.getTrackingNumberAvailableQty(stockLocation, trackingNumber);
+    }
+
+    StockLocationLine detailStockLocationLine =
+        Beans.get(StockLocationLineService.class)
+            .getDetailLocationLine(stockLocation, trackingNumber.getProduct(), trackingNumber);
+
+    BigDecimal availableQty = BigDecimal.ZERO;
+
+    if (detailStockLocationLine != null) {
+      availableQty =
+          detailStockLocationLine
+              .getCurrentQty()
+              .subtract(detailStockLocationLine.getReservedQty());
     }
     return availableQty;
   }